Refactor tool config conversion to avoid functions (4632)
