+++ clients/rust/src/lib.rs
@@ -1077,7 +1077,7 @@ impl ClientExt for Client {
         // Convert storage types to wire types
         let wire_inferences: Result<Vec<StoredInference>, _> = inferences
             .into_iter()
-            .map(|inf| inf.into_stored_inference(&gateway.handle.app_state.config))
+            .map(StoredInferenceDatabase::into_stored_inference)
             .collect();
 
         wire_inferences.map_err(|e| TensorZeroError::Other { source: e.into() })
