+++ evaluations/src/lib.rs
@@ -609,12 +609,10 @@ async fn infer_datapoint(params: InferDatapointParams<'_>) -> Result<InferenceRe
     };
 
     debug!("Processing tool parameters");
-    let function_config = config.get_function(function_name)?;
     let dynamic_tool_params = match datapoint.tool_call_config() {
         Some(tool_params) => {
             debug!("Tool parameters found, processing");
-            let function_config = config.get_function(function_name)?;
-            function_config.database_insert_to_dynamic_tool_params(tool_params.clone())
+            tool_params.clone().into()
         }
         None => {
             debug!("No tool parameters found");
@@ -622,6 +620,7 @@ async fn infer_datapoint(params: InferDatapointParams<'_>) -> Result<InferenceRe
         }
     };
     debug!("Processing output schema");
+    let function_config = config.get_function(function_name)?;
     let output_schema = match (datapoint.output_schema(), &**function_config) {
         // If the datapoint has an output schema, use it only in the case where it is not the same as the output schema of the function
         (Some(output_schema), FunctionConfig::Json(json_function_config)) => {
