+++ tensorzero-core/src/endpoints/datasets/v1/create_from_inferences.rs
@@ -10,7 +10,7 @@ use crate::db::inferences::{InferenceOutputSource, InferenceQueries, ListInferen
 use crate::endpoints::datasets::v1::types::CreateDatapointsFromInferenceOutputSource;
 use crate::endpoints::datasets::validate_dataset_name;
 use crate::error::{Error, ErrorDetails};
-use crate::stored_inference::StoredInference;
+use crate::stored_inference::{StoredInference, StoredInferenceDatabase};
 use crate::utils::gateway::{AppState, AppStateData, StructuredJson};
 
 use super::types::{
@@ -95,7 +95,7 @@ pub async fn create_from_inferences(
         .list_inferences(config, &list_inferences_params)
         .await?
         .into_iter()
-        .map(|x| x.into_stored_inference(config))
+        .map(StoredInferenceDatabase::into_stored_inference)
         .collect::<Result<Vec<_>, _>>()?;
 
     if let CreateDatapointsFromInferenceRequestParams::InferenceIds {
